Move STM (and clone) related UART implementation to platform
