+++ src/platform/STM32/platform_mcu.h
@@ -363,3 +363,100 @@ extern uint8_t _dmaram_end__;
 #if !defined(STM32G4) && !defined(STM32H7)
 #define USE_TX_IRQ_HANDLER
 #endif
+
+#if defined(STM32H7)
+#define UART_TX_BUFFER_ATTRIBUTE DMA_RAM            // D2 SRAM
+#define UART_RX_BUFFER_ATTRIBUTE DMA_RAM            // D2 SRAM
+#elif defined(STM32G4)
+#define UART_TX_BUFFER_ATTRIBUTE DMA_RAM_W          // SRAM MPU NOT_BUFFERABLE
+#define UART_RX_BUFFER_ATTRIBUTE DMA_RAM_R          // SRAM MPU NOT CACHABLE
+#elif defined(STM32F7)
+#define UART_TX_BUFFER_ATTRIBUTE FAST_DATA_ZERO_INIT // DTCM RAM
+#define UART_RX_BUFFER_ATTRIBUTE FAST_DATA_ZERO_INIT // DTCM RAM
+#elif defined(STM32F4)
+#define UART_TX_BUFFER_ATTRIBUTE                    // NONE
+#define UART_RX_BUFFER_ATTRIBUTE                    // NONE
+#endif
+
+#if defined(STM32F7) || defined(STM32H7) || defined(STM32G4)
+ // pin AF mode is configured for each pin individually
+#define UART_TRAIT_AF_PIN 1
+#else
+// all pins on given uart use same AF
+#define UART_TRAIT_AF_PORT 1
+#endif
+
+#if !defined(STM32F4)
+#define UART_TRAIT_PINSWAP 1
+#endif
+
+#if defined(STM32F4)
+
+#define UARTHARDWARE_MAX_PINS 4
+#ifndef UART_RX_BUFFER_SIZE
+#define UART_RX_BUFFER_SIZE     256
+#endif
+#ifndef UART_TX_BUFFER_SIZE
+#ifdef USE_MSP_DISPLAYPORT
+#define UART_TX_BUFFER_SIZE     1280
+#else
+#define UART_TX_BUFFER_SIZE     256
+#endif
+#endif
+
+#elif defined(STM32F7)
+
+#define UARTHARDWARE_MAX_PINS 4
+#ifndef UART_RX_BUFFER_SIZE
+#define UART_RX_BUFFER_SIZE     256
+#endif
+#ifndef UART_TX_BUFFER_SIZE
+#ifdef USE_MSP_DISPLAYPORT
+#define UART_TX_BUFFER_SIZE     1280
+#else
+#define UART_TX_BUFFER_SIZE     256
+#endif
+#endif
+
+#elif defined(STM32H7)
+
+#define UARTHARDWARE_MAX_PINS 5
+#ifndef UART_RX_BUFFER_SIZE
+#define UART_RX_BUFFER_SIZE     256
+#endif
+#ifndef UART_TX_BUFFER_SIZE
+#ifdef USE_MSP_DISPLAYPORT
+#define UART_TX_BUFFER_SIZE     1280
+#else
+#define UART_TX_BUFFER_SIZE     256
+#endif
+#endif
+
+#elif defined(STM32G4)
+
+#define UARTHARDWARE_MAX_PINS 3
+#ifndef UART_RX_BUFFER_SIZE
+#define UART_RX_BUFFER_SIZE     256
+#endif
+#ifndef UART_TX_BUFFER_SIZE
+#ifdef USE_MSP_DISPLAYPORT
+#define UART_TX_BUFFER_SIZE     1280
+#else
+#define UART_TX_BUFFER_SIZE     256
+#endif
+#endif
+
+#endif
+
+#if defined(STM32F7) || defined(STM32H7) || defined(STM32G4)
+#define UART_REG_RXD(base) ((base)->RDR)
+#define UART_REG_TXD(base) ((base)->TDR)
+#elif defined(STM32F4)
+#define UART_REG_RXD(base) ((base)->DR)
+#define UART_REG_TXD(base) ((base)->DR)
+#endif
+
+
+#if defined(STM32F4) || defined(STM32F7) || defined(STM32H7) || defined(STM32G4)
+#define DMA_TRAIT_CHANNEL
+#endif
